LayoutTests/imported/w3c/web-platform-tests/uievents/order-of-events/focus-events/focus-management-expectations-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Focus management from button  To button  End test button

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Keydown to focus should not trigger activation Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Keydown to focus should not trigger activation Test timed out
